1TE19248G Exercise 1 Submission 2
This is the final version of my submission, I did my best with the exercise. I think I did almost all of the tasks except for returning the error when the abstract method is not defined in the class that inherits it. Please let me know if there is anything I can improve on.